% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_liftover.R
\name{adaptive_liftover}
\alias{adaptive_liftover}
\title{adaptive_liftover}
\usage{
adaptive_liftover(
  gr,
  chain,
  window = 2000L,
  option = "adaptive",
  gap = 5L,
  step_size = 200L,
  verbose = TRUE
)
}
\arguments{
\item{gr}{The query regions of \code{\link[GenomicRanges]{GRanges}} class.}

\item{chain}{The chain object obtained from \code{rtracklayer::import.chain()}.}

\item{window}{The window size (bp) around the query regions. Default is 2000L.}

\item{option}{The option for adaptive_liftover
option = 'direct', apply direct liftOver to the query regions;
option = 'local', apply local liftOver by extending the query regions by a window size;
option = 'adaptive', apply direct liftOver first, for these failing to map, apply local liftOver.}

\item{gap}{Ignore the gaps among target regions if below this value (bp). Default it 5L.}

\item{step_size}{The step_size for generating candidate target regions. Default is 200L.}

\item{verbose}{Print messages or not. Default is TRUE.}
}
\value{
The candidate target regions of \code{\link[GenomicRanges]{GRangesList}} class with
length equal to the query regions.
}
\description{
Generate candidate target regions
}
\examples{

\dontrun{
data('data_example')
## the reciprocal chain file can be downloaded from UCSC
chain <- import.chain('mm10.hg38.rbest.chain')
gr_target_list <- adaptive_liftover(gr, chain, window = 2000, step_size = 10000)
gr_target_list <- adaptive_liftover(gr, chain, window = 2000, step_size = 200)
}


}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
