% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_grammar.R
\name{compute_grammar_matrix}
\alias{compute_grammar_matrix}
\title{Compute Grammar Matrix}
\usage{
compute_grammar_matrix(gr, hits_gr, all_motifs, grammar_size = 500L)
}
\arguments{
\item{gr}{GRanges object representing the regions.}

\item{hits_gr}{GRanges object of motif hits.}

\item{all_motifs}{Character vector of all motif names (human motifs).}

\item{grammar_size}{Integer, the size of the region to consider around each region.}
}
\value{
Sparse logical matrix indicating presence of motifs in regions.
}
\description{
This function builds the presence matrix of motifs for the given regions and hits data.
}
