% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_epigenome.R
\name{compute_similarity_epigenome}
\alias{compute_similarity_epigenome}
\title{compute_similarity_epigenome}
\usage{
compute_similarity_epigenome(
  gr_query,
  gr_target_list,
  query_grlist,
  target_grlist,
  weights = NULL,
  maxgap = 500L,
  metric = "cosine",
  verbose = TRUE
)
}
\arguments{
\item{gr_query}{The query genomic regions of \code{\link[GenomicRanges]{GRanges}} class.}

\item{gr_target_list}{The corresponding candidate target regions of \code{\link[GenomicRanges]{GRangesList}} class.
The length of gr_target_list should be the same as the length of gr_query}

\item{query_grlist}{The epigenomic datasets in query genome, \code{\link[GenomicRanges]{GRangesList}} class.}

\item{target_grlist}{The epigenomic datasets in target genome, \code{\link[GenomicRanges]{GRangesList}} class.
Should match with query_grlist.}

\item{weights}{The weights of epigenomic datasets. A numeric vector with length equal to the number of epigenomic datasets.
Default is NULL (equal weights).}

\item{maxgap}{The maxgap allowed for overlapping. Default is 500L.}

\item{metric}{The metric for computing similarities between indicators.
One of 'cosine' and 'jaccard'. Default is 'cosine'.}

\item{verbose}{Print messages or not. Default is TRUE.}
}
\value{
A \code{\link[GenomicRanges]{GRangesList}} class the same structure as gr_target_list
with an extra meta column named "epigenome" measuring the similarity score of epigenome
datasets.
}
\description{
This function computes the epigenome profile similarities between query regions and
the corresponding list of target regions.
}
\examples{

\dontrun{
data('data_example')
## the epigenomic datasets in this example can be downloaded from github
gr_target_list <- compute_similarity_epigenome(gr, gr_list, epigenome_mm10, epigenome_hg38)
}


}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
