% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_epigenome.R
\name{compute_similarity_epigenome_flat}
\alias{compute_similarity_epigenome_flat}
\title{compute_similarity_epigenome_flat}
\usage{
compute_similarity_epigenome_flat(
  gr_query,
  gr_target,
  query_grlist,
  target_grlist,
  weights = NULL,
  maxgap = 500L,
  metric = "cosine"
)
}
\arguments{
\item{gr_query}{The query genomic regions of \code{\link[GenomicRanges]{GRanges}} class.}

\item{gr_target}{The target genomic regions of \code{\link[GenomicRanges]{GRanges}} class.
The length of gr_target has to be equal to the length of gr_query.
Or either gr_query or gr_target has length 1 that can be broadcasted.}

\item{query_grlist}{The epigenomic datasets in query genome, \code{\link[GenomicRanges]{GRangesList}} class.}

\item{target_grlist}{The epigenomic datasets in target genome, \code{\link[GenomicRanges]{GRangesList}} class.
Should match with query_grlist.}

\item{weights}{The weights of epigenomic datasets. A numeric vector with length equal to the number of epigenomic datasets.
Default is NULL (equal weights).}

\item{maxgap}{The maxgap allowed for overlapping. Default is 500L.}

\item{metric}{The metric for computing similarities between indicators.
One of 'cosine' and 'jaccard'. Default is 'cosine'.}
}
\value{
A numeric vector with values between 0 and 1.
}
\description{
This function computes the epigenome profile similarities between query regions and
the corresponding target regions.
}
\examples{

\dontrun{
data('data_example')
## the epigenomic datasets in this example can be downloaded from github
similarity <- compute_similarity_epigenome_flat(gr[2], gr_list[[2]], epigenome_mm10, epigenome_hg38)
}

}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
