% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_grammar.R
\name{compute_similarity_from_matrix_grammar}
\alias{compute_similarity_from_matrix_grammar}
\title{Compute Similarity from Grammar Matrices}
\usage{
compute_similarity_from_matrix_grammar(
  query_mat,
  target_mat,
  metric = "cosine"
)
}
\arguments{
\item{query_mat}{Sparse matrix of query regions.}

\item{target_mat}{Sparse matrix of target regions.}

\item{metric}{Character, the similarity metric to use ('cosine' or 'jaccard').}
}
\value{
Numeric vector of similarity scores.
}
\description{
This function computes the similarity between two sets of regions based on their grammar matrices.
}
