% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_grammar.R
\name{compute_similarity_grammar}
\alias{compute_similarity_grammar}
\title{Compute Similarity Grammar}
\usage{
compute_similarity_grammar(
  gr_query,
  gr_target_list,
  hits_query_gr_list,
  hits_target_gr_list,
  motif_mapping,
  all_motifs,
  grammar_size = 500L,
  metric = "cosine",
  verbose = TRUE
)
}
\arguments{
\item{gr_query}{GRanges object representing the query regions.}

\item{gr_target_list}{GRangesList object representing the target regions for each query region.}

\item{hits_query_gr_list}{GRangesList object of motif hits for the query regions.}

\item{hits_target_gr_list}{GRangesList object of motif hits for the target regions.}

\item{motif_mapping}{Data frame containing the mapping between motifs in different species.
Should have two columns: 'mouse' and 'human'.}

\item{all_motifs}{Character vector of all motif names (human motifs).}

\item{grammar_size}{Integer, the size of the region to consider around each region.}

\item{metric}{Character, the similarity metric to use ('cosine' or 'jaccard').}

\item{verbose}{Logical, whether to print messages.}
}
\value{
GRangesList object with the same structure as gr_target_list, with an added
metadata column 'grammar' containing the similarity scores.
}
\description{
This function computes the sequence grammar similarities between query regions and
the corresponding list of target regions.
}
