% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_grammar.R
\name{compute_similarity_grammar_flat}
\alias{compute_similarity_grammar_flat}
\title{Compute Similarity Grammar Flat}
\usage{
compute_similarity_grammar_flat(
  gr_query,
  gr_target,
  hits_query_gr,
  hits_target_gr,
  motif_mapping,
  all_motifs,
  grammar_size = 500L,
  metric = "cosine"
)
}
\arguments{
\item{gr_query}{GRanges object representing the query regions.}

\item{gr_target}{GRanges object representing the target regions.}

\item{hits_query_gr}{GRanges object of motif hits for the query regions.}

\item{hits_target_gr}{GRanges object of motif hits for the target regions.}

\item{motif_mapping}{Data frame containing the mapping between motifs in different species.
Should have two columns: 'mouse' and 'human'.}

\item{all_motifs}{Character vector of all motif names (human motifs).}

\item{grammar_size}{Integer, the size of the region to consider around each region.}

\item{metric}{Character, the similarity metric to use ('cosine' or 'jaccard').}
}
\value{
Numeric vector of similarity scores.
}
\description{
This function computes the sequence grammar similarities between query regions and
the corresponding target regions.
}
