% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hits_query_gr_list.R
\name{generate_hits_query_gr_list}
\alias{generate_hits_query_gr_list}
\title{generate_hits_query_gr_list}
\usage{
generate_hits_query_gr_list(hits_query, gr_query)
}
\arguments{
\item{hits_query}{Data frame containing motif hits for the query species.
The data frame should have columns: 'chr', 'start', 'end', and 'motif_name'.
Can accept finemo hit calling result: hits.tsv.}

\item{gr_query}{A GRanges object representing the query regions.}
}
\value{
A GRangesList object where each element contains the motif hits for a corresponding query region.
}
\description{
This function generates a GRangesList object for the query regions by mapping motif hits.
}
