% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hits_target_gr_list.R
\name{generate_hits_target_gr_list}
\alias{generate_hits_target_gr_list}
\title{generate_hits_target_gr_list}
\usage{
generate_hits_target_gr_list(hits_target, gr_query, gr_target_list)
}
\arguments{
\item{hits_target}{Data frame containing motif hits for the target species.
The data frame should have columns: 'chr', 'start', 'end', and 'motif_name'.
Can accept finemo hit calling result: hits.tsv.}

\item{gr_query}{A GRanges object representing the query regions.}

\item{gr_target_list}{A GRangesList object representing the target regions for each query region.}
}
\value{
A list of GRangesList objects where each element corresponds to a query region and contains
GRanges objects for the hits in the associated target regions.
}
\description{
This function generates a nested list of GRangesList objects for the target regions by mapping motif hits.
}
