% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_candidate_filter.R
\name{gr_candidate_filter}
\alias{gr_candidate_filter}
\title{Filter Top Candidate Regions}
\usage{
gr_candidate_filter(
  gr_candidate,
  best_k = 1L,
  top_percentile = 0.01,
  random = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gr_candidate}{A \code{\link[GenomicRanges]{GRangesList}} object with a 'grammar' meta column.}

\item{best_k}{Integer, the number of top regions to select for each query region. Default is 1.}

\item{top_percentile}{Numeric, selects regions with grammar scores in the top percentile
(value between 0 and 1). Default is 0.01 (top 1\%).}

\item{random}{Logical, if TRUE, selects \code{best_k} regions randomly; otherwise, selects
the highest-scoring regions. Default is FALSE.}

\item{verbose}{Logical, whether to print messages. Default is TRUE.}
}
\value{
A \code{\link[GenomicRanges]{GRangesList}} object with filtered regions.
}
\description{
This function filters candidate target regions based on the top percentile
of grammar similarity scores.
}
