% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_grammar.R
\name{map_motifs}
\alias{map_motifs}
\title{Map Motifs Between Species}
\usage{
map_motifs(hits_gr, motif_mapping, from_species, to_species)
}
\arguments{
\item{hits_gr}{GRanges object containing motif hits with a 'motif' metadata column.}

\item{motif_mapping}{Data frame containing the mapping between motifs in different species.
Should have two columns: 'mouse' and 'human'.}

\item{from_species}{Character, the species of the motifs in hits_gr ('mouse' or 'human').}

\item{to_species}{Character, the species to map the motifs to ('mouse' or 'human').}
}
\value{
GRanges object with motifs mapped to the target species.
}
\description{
This function maps motifs from the query species to the target species using a provided mapping.
}
