% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_motif_list.R
\name{prepare_motif_list}
\alias{prepare_motif_list}
\title{prepare_motif_list}
\usage{
prepare_motif_list(meme_file, hits_file, motif_length = 30)
}
\arguments{
\item{meme_file}{Path to the PFM.meme file containing motif patterns}

\item{hits_file}{Path to the hits.tsv file containing hits calling results}

\item{motif_length}{Length of each motif (default: 30)}
}
\value{
A list containing:
- motif_list: A PWMatrixList object for use in Adaliftover
- hits_data: A GRanges object with motif mapping from hits.tsv
}
\description{
Prepare a motif list from chrombpnet/tf-modisco PFM.meme and finemo hits.tsv files
}
