% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_module.R
\name{training_module}
\alias{training_module}
\title{training_module}
\usage{
training_module(
  gr_candidate,
  gr_true,
  max_filter_proportion = 0.4,
  interaction = FALSE
)
}
\arguments{
\item{gr_candidate}{A \code{\link[GenomicRanges]{GRangesList}} object with 'epigenome' and 'grammar' meta columns.}

\item{gr_true}{A \code{\link[GenomicRanges]{GRanges}} object as the true orthologous genomic regions
that we aim to derive.}

\item{max_filter_proportion}{The maximum proportion of candidate target regions that are allowed to be filtered out. Default is 0.4.}

\item{interaction}{If we include the interaction term in the logistic regression. Default is FALSE.}
}
\value{
A data.table with the following columns:
\tabular{ll}{
\code{auroc} \tab The optimal AUROC. \cr
\code{aupr} \tab The optimal AUPR. \cr
\code{b.intercept} \tab The estimated regression coefficient for the intercept. \cr
\code{b.epigenome} \tab The estimated regression coefficient for the epigenome signal. \cr
\code{b.grammar} \tab The estimated regression coefficient for the sequence grammar. \cr
\code{b.interaction} \tab The estimated regression coefficient for the interaction term if included. \cr
\code{max_filter_proportion} \tab The input parameter max_filter_proportion. \cr
\code{threshold} \tab The suggested score threshold. \cr
\code{precision} \tab The precision associated with the suggested score threshold. \cr
}
}
\description{
This function takes GRangesList with 'epigenome' and 'grammar' meta columns
and evaluate the predictive power of overlapping with the true orthologous genomic regions.
}
\examples{

data('training_module_example')
res <- training_module(gr_candidate, gr_true)


}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
