% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_similarity_grammar.R
\name{compute_similarity_grammar}
\alias{compute_similarity_grammar}
\title{compute_similarity_grammar}
\usage{
compute_similarity_grammar(
  gr_query,
  gr_target_list,
  query_genome,
  target_genome,
  motif_list,
  grammar_size = 500L,
  metric = "cosine",
  verbose = TRUE
)
}
\arguments{
\item{gr_query}{The query genomic regions of \code{\link[GenomicRanges]{GRanges}} class.}

\item{gr_target_list}{The corresponding candidate target regions of \code{\link[GenomicRanges]{GRangesList}} class.
The length of gr_target_list should be the same as the length of gr_query}

\item{query_genome}{The query genome build identifier (e.g. 'mm10', 'hg38').
Make sure the corresponding BSgenome package is installed.
Or a \code{\link[BSgenome]{BSgenome}} class.}

\item{target_genome}{The target genome build identifier (e.g. 'mm10', 'hg38').
Make sure the corresponding BSgenome package is installed.
Or a \code{\link[BSgenome]{BSgenome}} class.}

\item{motif_list}{\code{\link[TFBSTools]{PWMatrixList}}, or \code{\link[TFBSTools]{PFMatrixList}}.}

\item{grammar_size}{The nearby region (bp) to extract regulatory grammars. Default is 500L.}

\item{metric}{The metric for computing similarities between indicators.
One of 'cosine' and 'jaccard'. Default is 'cosine'.}

\item{verbose}{Print messages or not. Default is TRUE.}
}
\value{
A \code{\link[GenomicRanges]{GRangesList}} class the same structure as gr_target_list
with an extra meta column named "grammar" measuring the similarity score of sequence grammars.
}
\description{
This function computes the sequence grammar similarities between query regions and
the corresponding list of target regions.
}
\examples{


data('data_example')
data('jaspar_pfm_list')
gr_target_list <- compute_similarity_grammar(gr, gr_list, 'mm10', 'hg38', jaspar_pfm_list)


}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
