% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_candidate_filter.R
\name{gr_candidate_filter}
\alias{gr_candidate_filter}
\title{gr_candidate_filter}
\usage{
gr_candidate_filter(
  gr_candidate,
  best_k = 1L,
  b_intercept = -4.1626668,
  b_epigenome = 5.4030573,
  b_grammar = 2.0649039,
  b_interaction = NULL,
  threshold = 0,
  random = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gr_candidate}{A \code{\link[GenomicRanges]{GRangesList}} object with 'epigenome' and 'grammar' meta columns.}

\item{best_k}{Select the best k regions for each element of gr_candidate. Default is 1L.}

\item{b_intercept}{The regression coefficient of intercept. Default is -4.1626668.}

\item{b_epigenome}{The regression coefficient of epigenome similarity. Default is 5.4030573.}

\item{b_grammar}{The regression coefficient of grammar similarity. Default is 2.0649039.}

\item{b_interaction}{The regression coefficient of the interaction term. Default is NULL.}

\item{threshold}{Discard the genomic regions with score below this threshold. Default is 0.}

\item{random}{Pick best_k regions randomly for each element of gr_candidate. Default is FALSE.}

\item{verbose}{Print messages or not. Default is TRUE.}
}
\value{
A \code{\link[GenomicRanges]{GRangesList}} class the same structure as gr_candidate
with an extra meta column named "score" measuring the overall similarity score.
Any genomic regions with score below the threshold will be filtered away.
}
\description{
This function takes GRangesList with 'epigenome' and 'grammar' meta columns
and filter the results based on these two columns.
}
\examples{

data('data_example')
gr_list_filtered <- gr_candidate_filter(gr_list_score, threshold = 0.5)


}
\author{
Chenyang Dong \email{cdong@stat.wisc.edu}
}
